重新生成代码

+++ TH.Mailer/Entity/EmailListobj.cs
@@ -8,16 +8,14 @@ using System.Text;
 using Pub.Class;
 
 namespace TH.Mailer.Entity {
- 
-    [Serializable]
-    [EntityInfo("")]
-    public partial class EmailListobj {
-        public IList<EmailList> list { get; set; }
-        public int page{ get; set; }
-        public int pages { get; set; }
-        public int totals { get; set; }
- 
-    }
+	[Serializable]
+	[EntityInfo("")]
+	public partial class EmailListobj {
+		public IList<EmailList> list { get; set; }
+		public int page { get; set; }
+		public int pages { get; set; }
+		public long totals { get; set; }
+	}
 }
